pyproject.toml
[tool.poetry]
name = "sf-permit-data"
version = "0.1.0"
description = ""
authors = ["goudete <goudetenrique@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.1"
python-dateutil = "^2.9.0.post0"
requests = "^2.31.0"
bs4 = "^0.0.2"
email-validator = "^2.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
